# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_controller.createAppReleaseReq:
    properties:
      auto_install_compose:
        type: boolean
      auto_install_docker:
        type: boolean
      cluster_id:
        type: integer
      cluster_name:
        type: string
      env_override:
        type: string
      install_dir:
        type: string
      name:
        type: string
      namespace:
        type: string
      project_name:
        type: string
      pull_images:
        type: boolean
      strategy:
        type: string
      target_id:
        type: integer
      target_name:
        type: string
      target_type:
        type: string
      template_id:
        type: integer
      values:
        additionalProperties: true
        type: object
    type: object
  internal_controller.createAppTemplateReq:
    properties:
      app_version:
        type: string
      category:
        type: string
      chart_name:
        type: string
      default_values:
        additionalProperties: true
        type: object
      engine:
        type: string
      env_example:
        type: string
      extra_files:
        items:
          type: string
        type: array
      install_dir_default:
        type: string
      manifest:
        type: string
      name:
        type: string
      owner:
        type: string
      project_name_default:
        type: string
      readme:
        type: string
      source:
        type: string
      source_ref:
        additionalProperties: true
        type: object
      source_type:
        type: string
      source_url:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      values_schema:
        additionalProperties: true
        type: object
      version:
        type: string
    type: object
...
  internal_controller.importRemoteAppTemplateReq:
    properties:
      category:
        type: string
      engine:
        type: string
      name:
        type: string
      source_type:
        type: string
      source_url:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  internal_controller.patchAppTemplateReq:
    properties:
      app_version:
        type: string
      category:
        type: string
      chart_name:
        type: string
      default_values:
        additionalProperties: true
        type: object
      engine:
        type: string
      env_example:
        type: string
      extra_files:
        items:
          type: string
        type: array
      install_dir_default:
        type: string
      manifest:
        type: string
      name:
        type: string
      owner:
        type: string
      project_name_default:
        type: string
      readme:
        type: string
      source:
        type: string
      source_ref:
        additionalProperties: true
        type: object
      source_type:
        type: string
      source_url:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      values_schema:
        additionalProperties: true
        type: object
      version:
        type: string
    type: object
...
  internal_controller.renderAppTemplateReq:
    properties:
      values:
        additionalProperties: true
        type: object
    type: object
...
  internal_controller.rollbackAppReleaseReq:
    properties:
      revision:
        type: integer
    type: object
...
  internal_controller.upgradeAppReleaseReq:
    properties:
      env_override:
        type: string
      pull_images:
        type: boolean
      values:
        additionalProperties: true
        type: object
    type: object
...
  k8s-platform-backend_pkg_resp.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 星枢K8S管理平台后端接口文档（统一响应结构：code/message/data）
  title: 星枢K8S管理平台 API
  version: "1.0"
paths:
  /api/v1/app-releases:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 集群 ID
        in: query
        name: cluster_id
        type: integer
      - description: 目标服务器 ID
        in: query
        name: target_id
        type: integer
      - description: 状态
        in: query
        name: status
        type: string
      - description: 模板 ID
        in: query
        name: template_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 应用发布列表
      tags:
      - apps
    post:
      parameters:
      - description: 发布信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.createAppReleaseReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 新建应用发布
      tags:
      - apps
  /api/v1/app-releases/{id}:
    delete:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 删除应用发布
      tags:
      - apps
    get:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 应用发布详情
      tags:
      - apps
  /api/v1/app-releases/{id}/revisions:
    get:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 应用发布版本历史
      tags:
      - apps
  /api/v1/app-releases/{id}/rollback:
    post:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 回滚参数
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_controller.rollbackAppReleaseReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 回滚应用发布
      tags:
      - apps
  /api/v1/app-releases/{id}/start:
    post:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 启动应用发布
      tags:
      - apps
  /api/v1/app-releases/{id}/stop:
    post:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 停止应用发布
      tags:
      - apps
  /api/v1/app-releases/{id}/toggle-pause:
    post:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 暂停/恢复应用发布
      tags:
      - apps
  /api/v1/app-releases/{id}/upgrade:
    post:
      parameters:
      - description: 发布 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 升级参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.upgradeAppReleaseReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 升级应用发布
      tags:
      - apps
  /api/v1/app-templates:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 分类
        in: query
        name: category
        type: string
      - description: 模板引擎
        in: query
        name: engine
        type: string
      - description: 模板状态
        in: query
        name: status
        type: string
      - description: 模板来源类型
        in: query
        name: source_type
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 应用模板列表
      tags:
      - apps
    post:
      parameters:
      - description: 模板信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.createAppTemplateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 新建应用模板
      tags:
      - apps
  /api/v1/app-templates/{id}:
    delete:
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 删除应用模板
      tags:
      - apps
    get:
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 应用模板详情
      tags:
      - apps
    patch:
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.patchAppTemplateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 更新应用模板
      tags:
      - apps
  /api/v1/app-templates/{id}/render:
    post:
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 渲染值
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.renderAppTemplateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 渲染应用模板预览
      tags:
      - apps
  /api/v1/app-templates/import-package:
    post:
      consumes:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 上传文件导入应用模板
      tags:
      - apps
  /api/v1/app-templates/import-remote:
    post:
      parameters:
      - description: 导入信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.importRemoteAppTemplateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/k8s-platform-backend_pkg_resp.Result'
      security:
      - BearerAuth: []
      summary: 从远程地址导入应用模板
      tags:
      - apps
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"